Issue 238 - Fix expiry times on status update records

+++ java/ingest/ingest-status-store/src/main/java/sleeper/ingest/status/store/job/DynamoDBIngestJobStatusFormat.java
@@ -58,10 +58,10 @@ public class DynamoDBIngestJobStatusFormat {
     public static final String UPDATE_TYPE_STARTED = "started";
     public static final String UPDATE_TYPE_FINISHED = "finished";
 
-    private final long timeToLiveInSeconds;
+    private final int timeToLiveInSeconds;
     private final Supplier<Instant> getTimeNow;
 
-    public DynamoDBIngestJobStatusFormat(long timeToLiveInSeconds, Supplier<Instant> getTimeNow) {
+    public DynamoDBIngestJobStatusFormat(int timeToLiveInSeconds, Supplier<Instant> getTimeNow) {
         this.timeToLiveInSeconds = timeToLiveInSeconds;
         this.getTimeNow = getTimeNow;
     }
